Release 1.0.14, more httpd changes.
httpd
 Add multi proc support
 Fix shutdown support
 Get rid of not finished 404 file support.
 Added configurations for:
  use_default_mime_type
  defer_accept
  mime_types_default_type
  mime_types_main
  mime_types_xtra
 Added loadable content type support
 Call abort() on a bad signal (allows core files).
 Fix escaping support for logging.
 Used a large reference for truncation data.
 Don't allow single \r or \n's in a path or host header.
 When at client max, stop accept() events (works better with multi. procs too).
 General cleanups and fixes, and more tests.

echod
 Fix shutdown support

evnt
 Fix shutdown support
 Allow evnt_*() to be called from evnt free callback.
 Fixups for valgrind errors.
 
ex_utils
 Have io_open() be blocking, and create io_open_nonblock().


git-archimport-id: james@and.org--2004-code/vstr--main--1--patch-86

+++ scripts/coverage.sh
@@ -49,7 +49,7 @@ function linkup()
   fi
 
 # Newer GCCs put them in the $srcdir
-  if [ ! -f ex_highlight-ex_highlight.da ]; then
+  if [ ! -f ex_highlight-ex_highlight.da -a ! -f vstr.da ]; then
     for i in .libs/*.da; do
       ln -f $i; rm -f $i
     done
